requirements.txt
Requests==2.32.5
schedule==1.2.2
websocket_client==1.9.0
